# Tweaks to the Dropwizard example