--- a/dropwizard/nologging.yml
+++ b/dropwizard/nologging.yml
@@ -4,7 +4,6 @@ server:
     - type: http
       port: 8080
       bindHost: 127.0.0.1 # only bind to loopback
-      inheritChannel: false
       headerCacheSize: 512 bytes
       outputBufferSize: 32KiB
       maxRequestHeaderSize: 8KiB
@@ -14,8 +13,6 @@ server:
       minBufferPoolSize: 64 bytes
       bufferPoolIncrement: 1KiB
       maxBufferPoolSize: 64KiB
-      acceptorThreads: 1
-      selectorThreads: 2
       acceptQueueSize: 1024
       reuseAddress: true
       soLingerTime: 345s
@@ -26,10 +23,10 @@ server:
     appenders: []
 
 logging:
-  level: WARN 
+  level: WARN
   appenders:
     - type: console
-      threshold: WARN 
+      threshold: WARN
       timeZone: UTC
       target: stdout
       logFormat: # TODO