--- a/dropwizard/src/main/java/com/mycompany/HelloWorldResource.java
+++ b/dropwizard/src/main/java/com/mycompany/HelloWorldResource.java
@@ -1,13 +1,14 @@
 package com.mycompany;
 
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
+import javax.inject.Singleton;
 
 @Path("/")
 @Produces(MediaType.TEXT_PLAIN)
+@Singleton
 public class HelloWorldResource {
     @GET
     public String sayHello() {