Mostly for my own comparison. I wanted to see how Dropwizard performs with a more realistic example which included the overhead of jersey's DI (HK2).

Also removed setting the acceptor/selector thread settings and letting it use the default. Controlling this variable makes sense, but I did not notice other benchmarks explicitly setting these values. 

Before (after 3 warmup runs)
```
proxy:microservices-framework-benchmark david$ wrk -t4 -c128 -d30s http://localhost:8080 -s pipeline.lua --latency -- / 16
Running 30s test @ http://localhost:8080
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   117.38ms  142.15ms   1.45s    85.50%
    Req/Sec     7.81k     3.42k   27.46k    75.75%
  Latency Distribution
     50%   58.06ms
     75%  154.14ms
     90%  321.41ms
     99%  639.80ms
  910336 requests in 30.03s, 86.82MB read
Requests/sec:  30316.34
Transfer/sec:      2.89MB
```

After (after 3 warmup runs)
```
proxy:microservices-framework-benchmark david$ wrk -t4 -c100 -d30s http://localhost:8080
Running 30s test @ http://localhost:8080
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.68ms   16.03ms 280.68ms   91.67%
    Req/Sec     5.96k     2.75k   20.38k    69.82%
  702814 requests in 30.07s, 67.03MB read
Requests/sec:  23369.43
Transfer/sec:      2.23MB
```